Fixed gets

+++ src/Repositories/IRepositoryInterface.cs
@@ -42,6 +42,6 @@ namespace NW.Repository
         );
 
         Task<Announcement> AddAnnouncement(Announcement announcement);
-        Task<Announcement[]> GetAnnouncements(bool? important, long fromTimestamp = 0, long toTimestamp = long.MaxValue);
+        Task<Announcement[]> GetAnnouncements(bool? important, long fromTimestamp = 0, long toTimestamp = int.MaxValue);
     }
 }
